+++ StudentEnrollingSystem/pom.xml
@@ -3,8 +3,8 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>StudentEnrollingSystem</groupId>
-	<artifactId>student-enrolling-system</artifactId>
+	<groupId>StudentEnrollmentSystem</groupId>
+	<artifactId>student-enrollment-system</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>com.student.entrolling</name>
