--- StudentEnrollingSystem/src/com/domain/enrolling/EnrollingApplication.java
@@ -1,5 +1,0 @@
-package com.domain.enrolling;
-
-public class EnrollingApplication {
-
-}
